+++ docs/source/changelog.rst
@@ -3,7 +3,7 @@ Changes
 
 0.1.0 [pre-alpha] 
 ------------------
-:Date: 2023-02-08
+:Date: 2023-02-12
 
 Contributors
 ~~~~~~~~~~~~
@@ -11,6 +11,7 @@ Contributors
 
 Fixes
 ~~~~~
+ - Add what's new to toctree.
  - Doc links broken.
  - Docs fetch the release numbers without import.
  - Docs don't attempt to load errors.
@@ -19,11 +20,14 @@ Fixes
 
 Improvements
 ~~~~~~~~~~~~
+ - Add interpolation tests.
  - Introduce `--build-num` so tests work again.
  - Better nox tests for skiylia build number.
 
 New features
 ~~~~~~~~~~~~
+ - String interpolation token + example.
+ - Lint the docs.
  - Add skiylia highlighter to documentation.
  - Introduce changelog and changelog workflow.